backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase
python-dotenv==1.0.0
httpx
stripe
